Add UDcide, an Android malware behavior editor
UDcide is a tool that provides alternative way to deal with Android malware. We help you to detect and remove specific behaviors in the malware rather than just delete the whole binary. And surprisingly, we make the binary runs still. This enables possibilities of malware research and makes good use of the normal behaviors in the malware.